+++ src/js/create-rating.js
