+++ src/js/filters.js
@@ -1,6 +1,6 @@
 import { showCategories } from './categories.js';
 import { ShowExercisesByCategory} from './exercises.js'
-import ExerciseFilterType from './exercise.filtertype.js';
+import ExerciseFilterType from './exercise.filter-type.js';
 
 const filterTabs = document.querySelectorAll('.filter-tab-button');
 const searchForm = document.querySelector('.filter-form-container');
